# backend/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:9090","description":"Generated server url"}],"paths":{"/stop":{"post":{"tags":["task-manager-controller"],"operationId":"stopThreads","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/start":{"post":{"tags":["task-manager-controller"],"operationId":"startThreads","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/save/vendor":{"post":{"tags":["vendor-controller"],"operationId":"saveVendor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vendor"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vendor"}}}}}}},"/save/ticket":{"post":{"tags":["ticket-controller"],"operationId":"saveTicket","parameters":[{"name":"ticket","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Ticket"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ticket"}}}}}}},"/save/customer":{"post":{"tags":["customer-controller"],"operationId":"saveCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/save/config":{"post":{"tags":["configuration-controller"],"operationId":"saveConfiguration","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Configuration"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Configuration"}}}}}}},"/reset":{"post":{"tags":["task-manager-controller"],"operationId":"resetLogs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/get/vendor":{"get":{"tags":["vendor-controller"],"operationId":"getVendors","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vendor"}}}}}}}},"/get/vendor/{vendorId}":{"get":{"tags":["vendor-controller"],"operationId":"getVendor","parameters":[{"name":"vendorId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vendor"}}}}}}},"/get/tickets":{"get":{"tags":["ticket-controller"],"operationId":"getTickets","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Ticket"}}}}}}}},"/get/customer":{"get":{"tags":["customer-controller"],"operationId":"getCustomers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}}}}}}},"/get/customer/{customerId}":{"get":{"tags":["customer-controller"],"operationId":"getCustomers_1","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/get/config":{"get":{"tags":["configuration-controller"],"operationId":"getConfiguration","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Configuration"}}}}}}},"/delete/vendor/{vendorId}":{"delete":{"tags":["vendor-controller"],"operationId":"deleteVendor","parameters":[{"name":"vendorId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/delete/customer":{"delete":{"tags":["customer-controller"],"operationId":"deleteCustomers","responses":{"200":{"description":"OK"}}}},"/delete/customer/{customerId}":{"delete":{"tags":["customer-controller"],"operationId":"deleteCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Vendor":{"type":"object","properties":{"vendorId":{"type":"integer","format":"int32"},"vendorName":{"type":"string"},"address":{"type":"string"},"email":{"type":"string"},"telNo":{"type":"string"},"eventName":{"type":"string"},"location":{"type":"string"},"ticketPrice":{"type":"number"},"ticketsPerRelease":{"type":"integer","format":"int32"}}},"Ticket":{"type":"object","properties":{"transactionId":{"type":"integer","format":"int32"},"ticketNo":{"type":"integer","format":"int32"},"vendor":{"type":"string"},"eventName":{"type":"string"},"location":{"type":"string"},"customer":{"type":"string"},"ticketPrice":{"type":"number"},"timestamp":{"type":"string","format":"date-time"}}},"Customer":{"type":"object","properties":{"customerId":{"type":"integer","format":"int32"},"customerName":{"type":"string"},"customerAddress":{"type":"string"},"customerEmail":{"type":"string"},"customerTel":{"type":"string"},"purchaseQuantity":{"type":"integer","format":"int32"}}},"Configuration":{"type":"object","properties":{"totalTickets":{"type":"integer","format":"int32"},"ticketReleaseRate":{"type":"number","format":"double"},"customerRetrievalRate":{"type":"number","format":"double"},"maxTicketCapacity":{"type":"integer","format":"int32"}}}}}}